set the config for adduser

+++ back-end/src/main/java/com/collegeApp/Back_end/model/User.java
@@ -14,8 +14,7 @@ import lombok.NoArgsConstructor;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private int id;
     private String username;
     private String password;
     private String role;
